Fix float truncation warning and suppress SFML C4275 warning
- App.cpp: change stepMs from int to float to avoid truncating 0.5 to 0
- MazeSolver-SFML.vcxproj: suppress C4275 (known SFML DLL interface warning)
- .gitignore: ignore entire .vscode/ directory (machine-specific config)

+++ MazeSolver-SFML/App.cpp
@@ -87,7 +87,7 @@ int main() {
     // Animation control variables
     // ---------------------------------------------------------------------
     sf::Clock clock;                  ///< Timer for controlling animation speed
-    const int stepMs = 0.5;           ///< Delay between animation steps (ms)
+    const float stepMs = 0.5f;         ///< Delay between animation steps (ms)
     std::size_t visitedIndex = 0;     ///< Index of current visited cell
     std::size_t pathIndex = 0;        ///< Index of current path cell
